refactor(workspace): extract URL state helpers
